pyproject.toml
[project]
name = "novel-writer-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "openai>=1.70.0",
    "pgvector>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.0",
]
